Fix issues in getMinPixels and getMaxPixels
The getMinPixels and getMaxPixels methods will never assert,
return null or return a negative value.

Remove assert statements in getMinPixels and getMaxPixels. The assert
in gitMinPixel was asserting if an application on an Android device
was run while the display was off because totalPixels was 0.

Handle totalPixels parameter being null.

Fixes #8